feat(codegen): cláusulas sobre coleção de aggregate carregado + projeção as V
I3.1 — caminho sem Collection[T]: quando o Target de um "load" é um
MemberExpr sobre a construção do aggregate (ex. "Wallet(id).entries"), a
lowering reusa LoadCall (E6.2, intocado) e aplica runtime.SelectSlice
sobre o campo de coleção do state — nenhum runtime.Collection[T]
envolvido, a fonte é a coleção já materializada. where/orderBy/skip/take
reusam INTEGRALMENTE o mesmo mecanismo de I1.1/I2.1 (hoistQueryPredicate/
buildLess), agora compartilhado entre o caminho Collection[T] (Policy/
Query via "list") e este novo caminho (aggregate carregado via "load
X(id).campo").

I3.2 — decl_query.go perde o fast-path que recusava cláusulas (E8.1): um
"return <QueryExpr com cláusulas>" delega ao hoisting de corpo normal. A
projeção "as V" reusa o mapeamento campo-a-campo com achatamento de VO
composto que "load X(id) as V" já tinha, aplicado por item num loop sobre
o resultado do SelectSlice/Select.

Âncora 1 (requirements §1.4): uma fixture Ledger (síntese, já que o
wallet real usa "list StatementEntry" sem cláusulas — não "load
Wallet(id).entries") prova a forma EXATA do GetStatement do spec §6.3
(load ... orderBy date descending skip page*20 take 20 as
StatementEntryVW) end-to-end.

Três bugs pré-existentes de import, achados e corrigidos ao verificar
(nunca exercitados antes deste ciclo, então nunca detectados):

1. skip/take: DS "integer" mapeia para Go int64, mas Query[T].Skip/Take
   são "int" — uma expressão como "page * 20" (page: int64) não compila
   sem cast; um literal como "20" compila por ser constante não-tipada.
   I2.1 só tinha exercitado literais; corrigido para emitir "int(...)"
   quando a expressão skip/take não é já um literal — as duas asserções
   de I2.1 que fixavam o texto sem o cast (orderby_test.go) foram
   atualizadas para refletir a forma correta.
2. decl_value.go: um ValueObject wrapper sobre "datetime" (goname.
   GoFieldType → "time.Time") gerava "type X time.Time" sem importar
   "time" — nenhum VO wrapper de datetime tinha golden/smoke test antes
   da fixture Ledger precisar de um campo "date" ordenável.
3. decl_view.go: um campo de View cujo tipo Go é "runtime.X" (ex.
   decimal → runtime.Decimal) não importava o runtime vendorado — nenhuma
   View real tinha campo decimal direto até a projeção "as V" (I3.2)
   achatar um Money para dentro de uma View.

Checkboxes de I3.1/I3.2 marcadas em tasks.md.

+++ codegen/lower/orderby_test.go
@@ -476,7 +476,7 @@ func TestHoistList_SkipTake_PlainExpressions(t *testing.T) {
 
 	out := lowerInFunc(t, l, StmtContext{}, "func testList()", assign)
 
-	want := `tmp1, err := tx.Select(ctx, runtime.Query[Money]{Skip: page * 20, Take: 20})`
+	want := `tmp1, err := tx.Select(ctx, runtime.Query[Money]{Skip: int(page * 20), Take: 20})`
 	if !strings.Contains(out, want) {
 		t.Fatalf("esperava %q, got:\n%s", want, out)
 	}
@@ -507,7 +507,7 @@ func TestHoistList_CompleteForm(t *testing.T) {
 	want := `tmp1, err := tx.Select(ctx, runtime.Query[StatementEntry]{` +
 		`Where: func(e StatementEntry) (bool, error) { return e.Description == TransactionDescription("Salário"), nil }, ` +
 		`Less: func(a, b StatementEntry) (bool, error) { return b.Description < a.Description, nil }, ` +
-		`OrderField: "description", OrderDesc: true, Skip: page * 20, Take: 20})`
+		`OrderField: "description", OrderDesc: true, Skip: int(page * 20), Take: 20})`
 	if !strings.Contains(out, want) {
 		t.Fatalf("esperava a Query[T] COMPLETA:\n%s\ngot:\n%s", want, out)
 	}
